Create auction and fetch by id

+++ Application/UseCases/Auction/Dto/DtoInputCreateAuction.cs
@@ -9,6 +9,6 @@ public class DtoInputCreateAuction
     [Required] public string Category { get; set; }
     [Required] public string Descri { get; set; }
     [Required] public string Img { get; set; }
-    [Required] public float Price { get; set; }
+    [Required] public decimal Price { get; set; }
     [Required] public DateTime Timer { get; set; } 
 }
